specialization/PythonForDataScience/requirements.txt
contourpy==1.3.1
cycler==0.12.1
flake8==7.1.1
fonttools==4.55.3
kiwisolver==1.4.7
matplotlib==3.10.0
mccabe==0.7.0
numpy==2.2.0
packaging==24.2
pillow==11.0.0
pycodestyle==2.12.1
pyflakes==3.2.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
six==1.17.0
